+++ chaospy/orthogonal/__init__.py
@@ -1,6 +1,0 @@
-r"""Collection of polynomial expansion constructors."""
-from .frontend import generate_expansion
-from .three_terms_recurrence import orth_ttr
-from .lagrange import lagrange_polynomial
-from .gram_schmidt import orth_gs
-from .cholesky import orth_chol